% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alpha_plots}
\alias{alpha_plots}
\title{Plot alpha trace and density}
\usage{
alpha_plots(alpha_posterior)
}
\arguments{
\item{alpha_posterior}{the posterior alpha object obtained by running the run_bmrr function}
}
\value{
a plot
}
\description{
Plot alpha trace and density
}
\examples{
iter = 1000
num_items = 10
variation = c(rep(0,iter), rep(.5, iter), rep(1, iter), rep(2, iter), rep(3, iter))
alpha_posterior <- array(
  abs(rnorm(iter*num_items, mean = 0 + variation, sd = 1)),
  dim = c(iter,num_items)
  )
 alpha_plots(alpha_posterior)
}
