% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{beta_plots}
\alias{beta_plots}
\title{Plot beta trace and density}
\usage{
beta_plots(beta_posterior, covariate_names = FALSE)
}
\arguments{
\item{beta_posterior}{the posterior beta object obtained by running the run_bmrr function}

\item{covariate_names}{a vector of covariate names that you want labeled on the graph}
}
\value{
a list of plots
}
\description{
Plot beta trace and density
}
\examples{
iter = 1000
num_covars = 5
variation = c(rep(-2,iter), rep(-1, iter), rep(0, iter), rep(1, iter), rep(2, iter))
beta_posterior <- array(
  rnorm(iter*num_covars, mean = 0 + variation, sd = 1),
  dim = c(iter,num_covars)
  )
covariate_names <- paste0("covariate_",1:num_covars)
beta_plots(beta_posterior, covariate_names)
}
