% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrr_data_sim.R
\name{bmrr_data_sim}
\alias{bmrr_data_sim}
\title{BMRR Sampler}
\usage{
bmrr_data_sim(
  P = 20,
  V = rep(10, P),
  H = 3,
  u = rep(0.5, P),
  nu = 0.5,
  cB = c(0, 0),
  cT = c(0, 0),
  cDA = 0,
  cDG = 0,
  s2T = 1,
  s2B = 1,
  s2A = 1,
  s2G = 1,
  s2DA = 1,
  s2DG = 1,
  covInd = c(TRUE, rep(FALSE, H - 1)),
  N = 25
)
}
\description{
Function to simulate data that can be used on the BMRR model by
the \link{bmrr_sampler} function. The data simulated follows the simulation
structure of the paper "Multi-Object Data Integration in the Study of
Primary Progressive Aphasia."
}
