% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrr_iterator.R
\name{bmrr_iterator}
\alias{bmrr_iterator}
\title{BMRR Iterator}
\usage{
bmrr_iterator(
  y,
  X,
  G,
  A,
  XG,
  Zv,
  Theta,
  B,
  DA,
  DG,
  C,
  t2T,
  l2T,
  vT,
  xiT,
  t2B,
  l2B,
  vB,
  xiB,
  sT2,
  sB2,
  nu,
  g,
  a_nu,
  b_nu,
  a_sT,
  b_sT,
  a_sB,
  b_sB
)
}
\arguments{
\item{y}{Main univariate variable of interest. Enter as a vector of size
\code{N} where \code{N} is the sample size.}

\item{X}{Covariates. Enter as a matrix of size \code{N} rows and \code{H} columns.
Where \code{H} is the number of not main covariates.}

\item{G}{Voxel object. Enter as an array of size (\code{N}, \code{mV}, \code{P}) where
\code{mV} is the maximum number of voxels on any region and \code{P} is
the number of regions. Use \code{NA} to fill the array if the number
of voxels per region is different or otherwise irregular, don't
use \code{0}.}

\item{A}{Network Object. Enter as an array of size (\code{N}, \code{P}, \code{P}).
Notice that the method doesn't take into consideration the
diagonal entries.}

\item{XG}{A replication of X into a matrix of \code{mV * N} rows by \code{H} columns.}

\item{Zv}{A matrization of the arrays \code{A} and \code{G} with \code{N} rows and
\code{P * (P - 1) / P + mV} columns equal to the number of
coefficients in both the Network object and the Voxel object.}

\item{Theta}{A matrix of size \code{P} by \code{P} corresponding to the
state in the MCMC of the Network coefficients.}

\item{B}{A matrix of size \code{mV} rows by \code{P} columns corresponding to the
state in the MCMC of the Voxel coefficients, where \code{mV} is the
maximum number of coefficients in a region.}

\item{DA}{A matrix of size \code{P} rows by \code{H} columns corresponding to the
state in the MCMC of the coefficients of the covariates in the
Network equations.}

\item{DG}{A matrix of size \code{P} rows by \code{H} columns corresponding to the
state in the MCMC of the of coefficients of the covariates in
the Voxels equations.}

\item{t2T}{A scalar corresponding to the state in the MCMC of the global
shrinking parameter of the Network coefficients prior.}

\item{l2T}{A matrix of size \code{P} by \code{P} corresponding to the state in the
MCMC of the local shrinking parameter of the Network
coefficients prior.}

\item{vT}{A matrix of size \code{P} by \code{P} corresponding to the state in the
MCMC of the auxiliary variable of the local shrinking parameter
of the Network coefficients prior.}

\item{xiT}{A scalar corresponding to the state in the MCMC of the auxiliary
variable corresponding to the global shrinking parameter of the
Network coefficients prior.}

\item{t2B}{A vector of size \code{P} corresponding to the state in the MCMC of
the global shrinking parameter for each region of the Voxel
coefficients prior.}

\item{l2B}{A matrix of size \code{mV} rows by \code{P} columns corresponding to the
state in the MCMC of the local shrinking parameter of the Voxel
coefficients prior.}

\item{vB}{A matrix of size \code{mV} rows by \code{P} columns corresponding to the
state in the MCMC of the auxiliary variable of the local
shrinking parameter of the Voxel coefficients prior.}

\item{xiB}{A vector of size \code{P} corresponding to the state in the MCMC of
the auxiliary variable corresponding to the global shrinking
parameter of each region of the Voxel coefficients prior.}

\item{sT2}{A scalar corresponding to the state in the MCMC of the variance
of the error in the Network equations.}

\item{sB2}{A scalar corresponding to the state in the MCMC of the variance
of the error in the Voxel equations.}

\item{nu}{An scalar for the probability of the Bernoulli prior on \code{g}.}

\item{g}{A binary vector of size \code{P}.}

\item{a_nu}{Shape 1 of the hyper_prior parameter for \code{nu}.}

\item{b_nu}{Shape 2 of the hyper_prior parameter for \code{nu}.}

\item{a_sT}{Shape of the hyper_prior parameter for \code{sT}.}

\item{b_sT}{Scale of the hyper_prior parameter for \code{sT}.}

\item{a_sB}{Shape of the hyper_prior parameter for \code{sB}.}

\item{b_sB}{Scale of the hyper_prior parameter for \code{sB}.}
}
\value{
A list containing all the updated parameters after one iteration of
the MCMC. That is, a list containing:
\itemize{
\item{\code{g}}{ A binary vector of size \code{P}.}
\item{\code{nu}}{An scalar for the probability of the Bernoulli prior on \code{g}.}
\item{\code{Theta}}{ A matrix of size \code{P} by \code{P} corresponding to the
state in the MCMC of the Network coefficients.}
\item{\code{B}}{ A matrix of size \code{mV} rows by \code{P} columns corresponding to the
state in the MCMC of the Voxel coefficients, where \code{mV} is the
maximum number of coefficients in a region.}
\item{\code{DA}}{ A matrix of size \code{P} rows by \code{H} columns corresponding to the
state in the MCMC of the coefficients of the covariates in the Network equations.}
\item{\code{DG}}{ A matrix of size \code{P} rows by \code{H} columns corresponding to the
state in the MCMC of the coefficients of the covariates in the Voxels equations.}
\item{\code{sT2}}{ A scalar corresponding to the state in the MCMC of the
variance of the error in the Network equations.}
\item{\code{sB2}}{ A scalar corresponding to the state in the MCMC of
the variance of the error in the Voxel equations.}
\item{\code{t2T}}{ A scalar corresponding to the state in the MCMC of
the global shrinking parameter of the Network coefficients prior.}
\item{\code{l2T}}{ A matrix of size \code{P} by \code{P} corresponding to the
state in the MCMC of the local shrinking parameter of the Network coefficients prior.}
\item{\code{xiT}}{ A scalar corresponding to the state in the MCMC of the auxiliary variable corresponding to the global shrinking parameter
of the Network coefficients prior.}
\item{\code{vT}}{ A matrix of size \code{P} by \code{P} corresponding to the
state in the MCMC of the auxiliary variable of the local shrinking parameter of the
Network coefficients prior.}
\item{\code{t2B}}{ A vector of size \code{P} corresponding to the state in the MCMC of the global
shrinking parameter of each region of the Voxel coefficients prior.}
\item{\code{l2B}}{ A matrix of size \code{mV} rows by \code{P} columns corresponding to the
state in the MCMC of the local shrinking parameter of the Voxel coefficients prior.}
\item{\code{xiB}}{ A vector of size \code{P} corresponding to the state in the MCMC of the auxiliary variable corresponding to the global shrinking parameter
of each region of the Voxel coefficients prior.}
\item{\code{vB}}{ A matrix of size \code{mV} rows by \code{P} columns corresponding to the
state in the MCMC of the auxiliary variable of the local shrinking parameter of the
Voxel coefficients prior.}
}
}
\description{
Performs an iteration of the MCMC for the Bayesian Multi-Object
Response Regression Method introduced in the paper "Multi-Object
Data Integration in the Study of Primary Progressive Aphasia."
}
