% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrr_sampler.R
\name{bmrr_sampler}
\alias{bmrr_sampler}
\title{BMRR Sampler}
\usage{
bmrr_sampler(
  y,
  X,
  G,
  A,
  nmcmc = 1000,
  burnin = 0,
  thining = 1,
  small_out = TRUE,
  state = list()
)
}
\arguments{
\item{y}{Main univariate variable of interest. Enter as a vector of
size \code{N} where \code{N} is the sample size.}

\item{X}{Covariates. Enter as a matrix of size \code{N} rows and \code{H}
columns. Where \code{H} is the number of not main covariates.}

\item{G}{Voxel object. Enter as an array of size (\code{N}, \code{mV}, \code{P})
where \code{mV} is the maximum number of voxels on any region and
\code{P} is the number of regions. Use \code{NA} to fill the array if
the number of voxels per region is different or otherwise
irregular, don't use \code{0}.}

\item{A}{Network Object. Enter as an array of size (\code{N}, \code{P}, \code{P}).
Notice that the method doesn't take into consideration the
diagonal entries.}

\item{nmcmc}{Number of MCMC samples. Default is \code{1000}.}

\item{burnin}{Number of samples to burn-in. By default there is no
burn-in, that is the parameter is set to \code{0}.}

\item{thining}{Natural number indicating how many iterations of the MCMC
have to be performed to obtain a sample. BY default there is
no thinning, that is the parameter is set to \code{1}.}

\item{small_out}{Parameter indicating if an output of small size is required.
Given the nature of problems to analyze, as in the paper,
the parameter space can be extremely big. If small_out is
used, the coefficient parameters are vectorized and stacked
for every iteration. Furthermore, the function doesn't
return second order parameters of method, that is only the
parameters at main regression functions and the region
indicators are returned.}

\item{state}{A list with all the parameters. It can be used to continue
the MCMC from a previous output or initialize the MCMC
chain. By default the algorithm doesn't require the
parameter list, and will initialize the method by itself.}
}
\value{
A list containing two elements:
\itemize{
\item{\code{sam}}{ A sample of the parameters, if \code{small_out = TRUE}:
\itemize{
\item{\code{g}}{ A binary matrix of size \code{nmcmc} rows and \code{P} columns.}
\item{\code{B}}{ A matrix of size \code{nmcmc} rows and \code{P * (P - 1) / P + mV}
columns , where \code{mV} is the number of voxels on the Voxel object,
corresponding to the samples of the coefficients. The order of the
coefficients is
obtained by stacking the coefficients of the network object and then the
coefficients for the voxel object. The coefficients of the network object
are ordered by \code{\link[=upper.tri]{upper.tri()}} and the coefficients of the voxel object are
ordered by stacking the columns of the array disregarding \code{NA}'s.}
\item{\code{DA}}{ An array of size \code{nmcmc} by \code{P} by \code{H} corresponding to the
samples of coefficients of the covariates in the Network equations.}
\item{\code{DG}}{ An array of size \code{nmcmc} by \code{P} by \code{H} corresponding to the
samples of coefficients of the covariates in the Voxels equations.}
\item{\code{sT2}}{ A vector of length \code{nmcmc} corresponding to the samples of
variance of the error in the Network equations.}
\item{\code{sB2}}{ A vector of length \code{nmcmc} corresponding to the samples of
the variance of the error in the Voxel equations.}
}
If \code{small_out = FALSE}:
\itemize{
\item{\code{g}}{ A binary matrix of size \code{nmcmc} rows and \code{P} columns.}
\item{\code{Theta}}{ An array of size \code{nmcmc} by \code{P} by \code{P} corresponding to the
samples of the Network coefficients.}
\item{\code{B}}{ An array of size \code{nmcmc} by \code{mV} by \code{P}  corresponding to the
samples of the Voxel coefficients, where \code{mV} is the maximum number of
coefficients in a region.}
\item{\code{DA}}{ An array of size \code{nmcmc} by \code{P} by \code{H} corresponding to the
samples of coefficients of the covariates in the Network equations.}
\item{\code{DG}}{ An array of size \code{nmcmc} by \code{P} by \code{H} corresponding to the
samples of coefficients of the covariates in the Voxels equations.}
\item{\code{sT2}}{ A vector of length \code{nmcmc} corresponding to the samples of
variance of the error in the Network equations.}
\item{\code{sB2}}{ A vector of length \code{nmcmc} corresponding to the samples of
the variance of the error in the Voxel equations.}
\item{\code{t2T}}{ A vector of length \code{nmcmc} corresponding to the samples of
the global shrinking parameter of the Network coefficients prior.}
\item{\code{l2T}}{ An array of size \code{nmcmc} by \code{P} by \code{P} corresponding to the
samples of the local shrinking parameter of the Network coefficients prior.}
\item{\code{xiT}}{ A vector of length \code{nmcmc} corresponding to the samples of
the auxiliary variable corresponding to the global shrinking parameter
of the Network coefficients prior.}
\item{\code{vT}}{ An array of size \code{nmcmc} by \code{P} by \code{P} corresponding to the
samples of the auxiliary variable of the local shrinking parameter of the
Network coefficients prior.}
\item{\code{t2B}}{ A vector of length \code{nmcmc} corresponding to the samples of
the global shrinking parameter of the Voxel coefficients prior.}
\item{\code{l2B}}{ An array of size \code{nmcmc} by \code{mV} rows by \code{P} columns corresponding to the
samples of the local shrinking parameter of the Voxel coefficients prior.}
\item{\code{xiB}}{ A vector of length \code{nmcmc} corresponding to the samples of
the auxiliary variable corresponding to the global shrinking parameter
of the Voxel coefficients prior.}
\item{\code{vB}}{ An array of size \code{nmcmc} by \code{mV} rows by \code{P} columns corresponding to the
samples of the auxiliary variable of the local shrinking parameter of the
Voxel coefficients prior.}
}
\item{\code{state}}{ The last state of the parameters of the MCMC chain. Can be
used to initialize another call to \code{\link[=bmrr_sample]{bmrr_sample()}}.}
}
}
\description{
Sampler for the Bayesian Multi-Object Response Regression
Method introduced in the paper "Multi-Object Data Integration in
the Study of Primary Progressive Aphasia." For details in the
sampling procedure computation see \code{\link[=bmrr_iterator]{bmrr_iterator()}}.
}
