% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_iterator.R
\name{group_iterator}
\alias{group_iterator}
\title{BMRR Iterator}
\usage{
group_iterator(Say, Sgy, Syy, sT2, sB2, LT, LB, g, nu, C, p)
}
\arguments{
\item{Say}{The sum (over observations) of the product of the main variable
\code{y} and the difference of the Network values and the effect of the
covariates. Input as a matrix of \code{P} by \code{P}.}

\item{Sgy}{The sum (over observations) of the product of the main variable
\code{y} and the difference of the Voxel values and the effect of the
covariates. Input as a matrix of \code{mV} rows and \code{P} columns.}

\item{Syy}{The sum (over observations) of the squares of the main variable
\code{y}.}

\item{sT2}{Variance of the error term of the Network equations.}

\item{sB2}{Variance of the error term of the Voxel equations.}

\item{LT}{Product of the global and shrinkage variables for the Network
coefficients. Input as a matrix of \code{P} by \code{P}.}

\item{LB}{Product of the global and shrinkage variables for the Voxel
coefficients.  Input as a matrix of \code{mV} rows and \code{P} columns.}

\item{g}{Binary indicators for every region.}

\item{C}{Boolean matrix for the Voxel structure.}

\item{p}{Region of interest.}
}
\description{
Performs the joint sampling of the indicator variable, the
Voxel coefficients and the Network coefficients for one region.
}
