% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mu_beta_plots}
\alias{mu_beta_plots}
\title{Plot mu + betaX trace and density}
\usage{
mu_beta_plots(
  mu_posterior,
  beta_posterior,
  covars,
  item_names = FALSE,
  criteria_names = FALSE
)
}
\arguments{
\item{mu_posterior}{the posterior mu object obtained by running the run_bmrr function}

\item{beta_posterior}{the posterior beta object obtained by running the run_bmrr function}

\item{covars}{the covariate data array}

\item{item_names}{a vector of items names that you want labeled on the graph}

\item{criteria_names}{a vector of criteria names that you want labeled on the graph}
}
\value{
a list of plots
}
\description{
Plot mu + betaX trace and density
}
\examples{
iter = 1000
num_items = 10
num_criteria = 3
num_covars = 5
variation = c(rep(-2,iter), rep(-1, iter), rep(0, iter), rep(1, iter), rep(2, iter))
mu_posterior <- array(
  rnorm(iter*num_items*num_criteria, mean = 0 + variation, sd = 1),
  dim = c(iter,num_items,num_criteria)
)

beta_posterior <- array(
  rnorm(iter*num_covars, mean = 0 + variation, sd = 1),
  dim = c(iter,num_covars)
)

covars <- array(
  rnorm(num_items*num_covars*num_criteria, mean = 0, sd = 1),
  dim = c(num_items, num_covars, num_criteria)
)



item_names <- paste("player_",1:num_items)
criteria_names <- paste("time_",1:num_criteria)
mu_beta_plots(mu_posterior, beta_posterior, covars, item_names, criteria_names)
}
