% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mu_plots}
\alias{mu_plots}
\title{Plot mu trace and density}
\usage{
mu_plots(mu_posterior, item_names = FALSE, criteria_names = FALSE)
}
\arguments{
\item{mu_posterior}{the posterior mu object obtained by running the run_bmrr function}

\item{item_names}{a vector of items names that you want labeled on the graph}

\item{criteria_names}{a vector of criteria names that you want labeled on the graph}
}
\value{
a list of plots
}
\description{
Plot mu trace and density
}
\examples{
iter = 1000
num_items = 10
num_criteria = 3
 variation = c(rep(-2,iter), rep(-1, iter), rep(0, iter), rep(1, iter), rep(2, iter))
mu_posterior <- array(
  rnorm(iter*num_items*num_criteria, mean = 0 + variation, sd = 1),
  dim = c(iter,num_items,num_criteria)
  )
item_names <- paste0("player_",1:num_items)
criteria_names <- paste0("time_",1:num_criteria)
 mu_plots(mu_posterior, item_names, criteria_names)
}
