% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrr.R
\name{run_bmrr}
\alias{run_bmrr}
\title{Run Bayesian Multivariate Rank Regression Model}
\usage{
run_bmrr(
  num_raters,
  num_items,
  num_criteria,
  data,
  samps,
  burnin,
  covars_present,
  covars,
  mu_0,
  sigma_0,
  alpha_0,
  beta_0,
  sum_0_inverse,
  lambda,
  seed
)
}
\arguments{
\item{num_raters}{number of raters: integer}

\item{num_items}{number of items: integer}

\item{num_criteria}{number of criteria: integer}

\item{data}{the data: array}

\item{samps}{number of samples post-burnin: integer}

\item{burnin}{number of burnin samples: integer}

\item{covars_present}{TRUE if covariates, FALSE if none}

\item{covars}{covariate data: array}

\item{mu_0}{prior mean for xi: real number}

\item{sigma_0}{prior variance for xi: real number}

\item{alpha_0}{prior for alpha: real number}

\item{beta_0}{prior mean for beta: real number}

\item{sum_0_inverse}{prior variance for beta: real number}

\item{lambda}{prior for sigma: real number}

\item{seed}{it's a seed what more do you need: integer}
}
\value{
list of posterior values with burnin removed
}
\description{
Run Bayesian Multivariate Rank Regression Model
}
\examples{

num_raters <- 10

}
