% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sigma_plots}
\alias{sigma_plots}
\title{Plot sigma  density}
\usage{
sigma_plots(sigma_posterior, ranker_names)
}
\arguments{
\item{sigma_posterior}{the posterior sigma object obtained by running the run_bmrr function}

\item{ranker_names}{a vector of ranker names that you want labeled on the graph}
}
\value{
a list of plots
}
\description{
Plot sigma  density
}
\examples{
iter = 1000
num_raters = 8
variation = c(rep(0,iter), rep(.5, iter), rep(1, iter), rep(2, iter))
sigma_posterior <- array(
  abs(rnorm(iter*num_raters, mean = 0 + variation, sd = 1)),
  dim = c(iter,num_raters)
  )
ranker_names <- paste0("rater_",1:num_raters)
sigma_plots(sigma_posterior, ranker_names)
}
