% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports_data.R
\docType{data}
\name{sports_data}
\alias{sports_data}
\title{Ranking of top MLB players from various new sources with covariates}
\format{
Data frame with 22 columns and 1127 rows
\describe{
\item{Player}{player name.}
\item{Position}{player position.}
\item{Year}{year ranked.}
\item{ESPN}{rank by ESPN for year.}
\item{CBS}{rank by CBS for year.}
\item{MLB}{rank by MLB for year.}
\item{BR}{rank by Bleacher Report for year.}
\item{Yahoo}{rank by Yahoo for year.}
\item{player_id}{unique player id.}
\item{player_age}{player age at year.}
\item{WAR}{Wins Above Replacement from previous year.}
\item{BaseSalary}{Base salary from previous year.}
}
}
\source{
\url{https://baseballsavant.mlb.com/statcast_search}
}
\usage{
sports_data
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  sports_data
}
\keyword{datasets}
