% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{xi_plots}
\alias{xi_plots}
\title{Plot xi trace and density}
\usage{
xi_plots(xi_posterior, item_names = FALSE)
}
\arguments{
\item{xi_posterior}{the posterior xi object obtained by running the run_bmrr function}

\item{item_names}{a vector of items names that you want labeled on the graph}
}
\value{
a list of plots
}
\description{
Plot xi trace and density
}
\examples{
iter = 1000
num_items = 10
variation = c(rep(-2,iter), rep(-1, iter), rep(0, iter), rep(1, iter), rep(2, iter))
xi_posterior <- array(
  rnorm(iter*num_items, mean = 0 + variation, sd = 1),
  dim = c(iter,num_items)
  )
item_names <- paste0("player_",1:num_items)
 xi_plots(xi_posterior, item_names)
}
